requirements.txt
streamlit
google-generativeai
python-dotenv
psycopg2